feature/event-form-refactor: replace types into separate folder
