+++ src/services/utils.ts
@@ -1,7 +1,7 @@
 import classNames, {ArgumentArray} from 'classnames';
 import {twMerge} from 'tailwind-merge';
 
-import {StorageKeys} from './types';
+import {StorageKeys} from '../types/types';
 
 export const getLocalStoredValues = (key: StorageKeys, defaultValues?: unknown) => {
   const storedValues = localStorage.getItem(key);
