+++ src/types/eventFormTypes.ts
@@ -1,0 +1,4 @@
+export enum FormActionType {
+  edit = 'edit',
+  create = 'create',
+}
